Fix POSTGRES_DB variable name and usage
This variable must be POSTGRES_DB for the postgres docker container.

It was changed to POSTGRES_DBNAME in 487e849e to match the
docker-compose scripts, which were incorrect.

The scripts were also incorrectly using POSTGRES_USER as the database
name, which is the default behavior of the container if POSTGRES_DB is
not specified ... so it accidentally worked if they were the same.